requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
apscheduler==3.10.4
python-dotenv==1.0.0
pyyaml==6.0.1
openpyxl==3.1.2
pandas==2.1.4
requests==2.31.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
aiohttp==3.9.1